requirements.txt
fastapi==0.86.0
Pillow==9.3.0
pydantic==1.10.2
python-multipart==0.0.5
uvicorn==0.19.0
